backend/requirements.txt
flask
flask-cors
pytesseract
pillow
PyPDF2
pdfplumber
requests
opencv-python-headless
numpy